refactor: getSavedRecommend 함수 이동
