+++ src/components/recommendStar/recommendStar.ts
@@ -1,7 +1,8 @@
 import filledStar from '../../images/star_filled.png';
 import emptyStar from '../../images/star_empty.png';
 import { SCORE_MESSAGE } from '../../constants/constant';
-import { UserMovieDetail } from '../../interface/Movie';
+import { MovieDetail, UserMovieDetail } from '../../interface/Movie';
+import { MovieDetailToUserMovieDetail } from '../../domain/MovieService';
 export class RecommendStar {
   private userMovieDetail;
   private currentScore: number;
@@ -89,7 +90,18 @@ export class RecommendStar {
       this.userMovieDetail.userVote = this.currentScore;
       recommendList.push(this.userMovieDetail);
     }
-    console.log(recommendList);
     localStorage.setItem('recommendList', JSON.stringify(recommendList));
   }
+}
+
+export function getSavedRecommend(movieDetail: MovieDetail) {
+  const recommendList = JSON.parse(localStorage.getItem('recommendList') || '[]');
+  if (recommendList.length !== 0) {
+    const userMovieDetail = recommendList.find((movie: UserMovieDetail) => movie.id === movieDetail.id);
+    if (userMovieDetail) {
+      return userMovieDetail;
+    }
+    return MovieDetailToUserMovieDetail(movieDetail, 0);
+  }
+  return MovieDetailToUserMovieDetail(movieDetail, 0);
 }
